Bug 1698097: Null check parent content when checking whether a tr should use ARIARowAccessible. r?Morgan
We assumed parent content wasn't null because we've already checked that there is a parent table accessible, suggesting there's also parent content.
However, if the tr is at the top level of a shadow root (but the table is not), parent content will be null, causing a crash.
Before the fix for bug 1686123, this was fine because the frame for a shadow root is null and we didn't continue for a null frame.
Now that we do continue for a null frame, we must null check parent content.

Differential Revision: https://phabricator.services.mozilla.com/D108777