Bug 1824064 abandoned: Don't cycle collect AccEvent.
While there can be cycles between AccEvents (AccTreeMutationEvent::mNext/PrevEvent), we know how and when to reliably break these.
Thus, we can avoid some overhead by not having them participate in cycle collection.
Note that this patch is not complete.
For example, if we return early in ProcessMutationEvents or if we shut down NotificationController before processing mutation events, there will be unbroken cycles.
We would need to explicitly deal with those cases.
I also haven't been able to prove that this has any observable benefit, so this is just a patch for reference in case it's useful in future.

Differential Revision: https://phabricator.services.mozilla.com/D173505