Use CoreAnimation Compositing on iOS
This brings the graphics and compositing setup more in line with the setup
currently used on macOS, making it possible for us to actually render content
to the screen with WebRender.

There is likely more painting code form the old setup which is now dead and can
be cleaned up.