Use correctly sized stacks on iOS
On iOS devices, the stack size must be a multiple of the page size. This
conflicts with some stacks which are sized to be 2MB-8192 bytes, which is not a
multiple of the 16k page size on iOS.

This patch hackily makes them a multiple of the page size on iOS as well.