Bug 1859357 - Part 1: Expose a basic GeckoView-like API for iOS embedding, r=glandium!
This is an initial patch doing a re-work of the way that the GeckoView windows
are created, sized, etc. The goal was to make the system more similar with
GeckoView on Android, and expose a pure Objective-C embedding API, which can be
used from Swift, similar to the Java API exposed by GeckoView on Android.

At the moment, the implementation of this API is very temporary and will need
further iteration and unification with the Android implementation.

This change breaks the ExampleBrowser project in the tree, as it was using the
old embedding strategy. That project will be replaced in the next part.

Differential Revision: https://phabricator.services.mozilla.com/D191100