Bug 1859357 - Part 3: Add basic ./mach run and ./mach install support for iOS, r=glandium!
These patches add extremely-basic implementations of ./mach run and ./mach
install using the TestRunner application added in the previous part.

When running against the simulator, these will use the `simctl` command to
install the application to the simulator and run it, and will use the
third-party `ios-deploy` tool to install and run the application on a real
device. I have so far struggled to find an officially-supported way to install
and run an iOS xcode project on device from the command line.

We may be able to use a similar approach to run real test suites if we're
comfortable with the ios-deploy dependency, though I expect significant changes
will need to be made to make the building and bundling setup compatible with
cross-compilation and CI.

This mostly acts as a proof-of-concept which aims to be useful for onboarding
new folks into working with Gecko on iOS.

Differential Revision: https://phabricator.services.mozilla.com/D191103