Bug 1837126: Tweak Android a11y testClipboard so that the results are consistent regardless of when events are fired. r?eeejay
When pasting replaces text, a11y might handle the removal and the insertion in the same tick or in two separate ticks.
This impacts the events that are fired because TextUpdater will diff the text (searching for unchanged characters) if the events occur in the same tick.
Also, the text might not be up to date yet in the parent process if the events are fired in separate ticks.
To deal with this, modify the text and selection so that there are no common characters between the text being replaced and the text it is replaced with, and assert the final text when the insertion event is fired.
See the comments in the test for more details.

Differential Revision: https://phabricator.services.mozilla.com/D225569